Change winnings type from int to double

+++ lab4/RaceCarDriver.java
@@ -2,10 +2,10 @@ package lab4;
 
 public class RaceCarDriver extends Person implements Raceable,Comparable<Raceable> {
     private int starts;
-    private int winnings;
+    private double winnings;
     private int wins;
 
-    public RaceCarDriver(String name, String country, int yearBorn, int starts, int winnings, int wins) {
+    public RaceCarDriver(String name, String country, int yearBorn, int starts, double winnings, int wins) {
         super(name, country, yearBorn);
 
         this.starts = starts;
@@ -13,7 +13,7 @@ public class RaceCarDriver extends Person implements Raceable,Comparable<Raceabl
         this.wins = wins;
     }
 
-    public int getWinnings() {
+    public double getWinnings() {
         return winnings;
     }
